manual: Clarify which structure editing commands use the region
* doc/org-manual.org (Structure Editing): Combine org-do-promote and
org-do-demote into a single entry that includes the remark about their
treatment of an active region.

At the end of the structure editing node, there is a note that
promotion and demotion works on the region, if active, but that
applies only to org-do-demote and org-do-promote, not
org-promote-subtree and org-demote-subtree.  Reposition this comment
to avoid any confusion.

Suggested-by: B Goodr <bgoodr@gmail.com>