manual: Clarify behavior of :file-desc
* doc/org-manual.org (Results of Evaluation): Distinguish between
behavior when the :file-desc header argument is not present and when
it has no value.

The current handling of the :file-desc argument is to omit the
description if the header argument isn't present and to use the :file
value as the description if :file-desc is present without a value.
This matches the original intention [*].

[*] https://orgmode.org/list/87vclky211.fsf@med.uni-goettingen.de

Reported-by: Matt Huszagh <huszaghmatt@gmail.com>
  https://orgmode.org/list/87sgbzl6uy.fsf@gmail.com